# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Primax Gas Distributor Order Management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  in_transit: 'in_transit',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  cash: 'cash',
  yape_plin: 'yape_plin',
  pos_card: 'pos_card',
} as const;

export interface Customer {
  id: number;
  name: string;
  phone: string;
  address: string;
  reference: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomerInput {
  name: string;
  phone: string;
  address: string;
  reference: string;
}

export interface UpdateCustomerInput {
  name?: string;
  phone?: string;
  address?: string;
  reference?: string;
}

export type DriverStatus = typeof DriverStatus[keyof typeof DriverStatus];


export const DriverStatus = {
  available: 'available',
  busy: 'busy',
  offline: 'offline',
} as const;

export interface Driver {
  id: number;
  name: string;
  phone: string;
  status: DriverStatus;
  latitude?: number | null;
  longitude?: number | null;
  locationUpdatedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Order {
  id: number;
  customer: Customer;
  product: string;
  /** @minimum 1 */
  quantity: number;
  paymentMethod: PaymentMethod;
  cashAmount?: number | null;
  totalAmount: number;
  status: OrderStatus;
  driverId?: number | null;
  driver?: Driver | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderInput {
  customerId?: number | null;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerReference: string;
  driverId?: number | null;
  product: string;
  /** @minimum 1 */
  quantity: number;
  paymentMethod: PaymentMethod;
  cashAmount?: number | null;
  totalAmount: number;
  notes?: string | null;
}

export interface UpdateOrderStatusInput {
  status: OrderStatus;
}

export interface OrderUpdate {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerReference: string;
  product: string;
  /** @minimum 1 */
  quantity: number;
  paymentMethod: PaymentMethod;
  /** @nullable */
  cashAmount?: number | null;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  driverId?: number | null;
  status?: OrderStatus;
}

export interface CreateDriverInput {
  name: string;
  phone: string;
  status?: DriverStatus;
}

export interface DriverUpdate {
  name?: string;
  phone?: string;
  status?: DriverStatus;
}

export interface DriverLocationUpdate {
  latitude: number;
  longitude: number;
}

export interface Settlement {
  date: string;
  driverId: number | null;
  driverName: string;
  cash: number;
  yapePlin: number;
  posCard: number;
  grandTotal: number;
  orders: number;
}

export interface TodaySummary {
  totalOrders: number;
  pending: number;
  inTransit: number;
  delivered: number;
  cancelled: number;
  totalRevenue: number;
}

export type ListCustomersParams = {
/**
 * Search by name or phone
 */
q?: string;
};

export type ListOrdersParams = {
/**
 * Filter by date (YYYY-MM-DD), defaults to today
 */
date?: string;
/**
 * Filter by status
 */
status?: OrderStatus;
};

export type GetDriverSettlementParams = {
date: string;
driverId?: number;
};