# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Primax Gas Distributor Order Management
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: customers
    description: Customer management
  - name: orders
    description: Order management

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: q
          in: query
          description: Search by name or phone
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerInput"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: date
          in: query
          description: Filter by date (YYYY-MM-DD), defaults to today
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Edit an order and its customer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/summary/today:
    get:
      operationId: getTodaySummary
      tags: [orders]
      summary: Get today's summary statistics
      responses:
        "200":
          description: Today's summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodaySummary"

  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List delivery drivers
      responses:
        "200":
          description: Delivery drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a delivery driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverInput"
      responses:
        "201":
          description: Created driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /drivers/{id}:
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a delivery driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /drivers/{id}/location:
    patch:
      operationId: updateDriverLocation
      tags: [drivers]
      summary: Update the driver's current location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverLocationUpdate"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  /settlements:
    get:
      operationId: getDriverSettlement
      tags: [drivers]
      summary: Get driver payment settlement
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: driverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Driver settlement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OrderStatus:
      type: string
      enum: [pending, in_transit, delivered, cancelled]

    PaymentMethod:
      type: string
      enum: [cash, yape_plin, pos_card]

    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        reference:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - address
        - reference
        - createdAt
        - updatedAt

    CreateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        reference:
          type: string
      required:
        - name
        - phone
        - address
        - reference

    UpdateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        reference:
          type: string

    Order:
      type: object
      properties:
        id:
          type: integer
        customer:
          $ref: "#/components/schemas/Customer"
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        cashAmount:
          type: number
          nullable: true
        totalAmount:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        driverId:
          type: integer
          nullable: true
        driver:
          allOf:
            - $ref: "#/components/schemas/Driver"
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customer
        - product
        - quantity
        - paymentMethod
        - totalAmount
        - status
        - createdAt
        - updatedAt

    CreateOrderInput:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerReference:
          type: string
        driverId:
          type: integer
          nullable: true
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        cashAmount:
          type: number
          nullable: true
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - customerName
        - customerPhone
        - customerAddress
        - customerReference
        - product
        - quantity
        - paymentMethod
        - totalAmount

    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - status

    OrderUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        customerReference:
          type: string
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        cashAmount:
          type: ["number", "null"]
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        driverId:
          type: integer
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - customerName
        - customerPhone
        - customerAddress
        - customerReference
        - product
        - quantity
        - paymentMethod
        - totalAmount

    DriverStatus:
      type: string
      enum: [available, busy, offline]

    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/DriverStatus"
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        locationUpdatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - status
        - createdAt
        - updatedAt

    CreateDriverInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/DriverStatus"
      required:
        - name
        - phone

    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/DriverStatus"

    DriverLocationUpdate:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude

    Settlement:
      type: object
      properties:
        date:
          type: string
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
        cash:
          type: number
        yapePlin:
          type: number
        posCard:
          type: number
        grandTotal:
          type: number
        orders:
          type: integer
      required:
        - date
        - driverId
        - driverName
        - cash
        - yapePlin
        - posCard
        - grandTotal
        - orders

    TodaySummary:
      type: object
      properties:
        totalOrders:
          type: integer
        pending:
          type: integer
        inTransit:
          type: integer
        delivered:
          type: integer
        cancelled:
          type: integer
        totalRevenue:
          type: number
      required:
        - totalOrders
        - pending
        - inTransit
        - delivered
        - cancelled
        - totalRevenue