# lib/api-zod/src/generated/types/createDriverInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Primax Gas Distributor Order Management
 * OpenAPI spec version: 0.1.0
 */
import type { DriverStatus } from './driverStatus';

export interface CreateDriverInput {
  name: string;
  phone: string;
  status?: DriverStatus;
}