# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Primax Gas Distributor Order Management
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from './paymentMethod';

export interface CreateOrderInput {
  customerId?: number | null;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerReference: string;
  driverId?: number | null;
  product: string;
  /** @minimum 1 */
  quantity: number;
  paymentMethod: PaymentMethod;
  cashAmount?: number | null;
  totalAmount: number;
  notes?: string | null;
}