# lib/api-zod/src/generated/types/driverLocationUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Primax Gas Distributor Order Management
 * OpenAPI spec version: 0.1.0
 */

export interface DriverLocationUpdate {
  latitude: number;
  longitude: number;
}